nan/requirements.txt
anyjson==0.3.3
Django==2.2.6
django-admin-generator==2.2.0
django-admin-interface==0.11.2
django-colorfield==0.1.15
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-utils2==2.5.4
python-utils==2.3.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
pillow
djangorestframework